// .PRD/erp-refactor/v3/types/subcontracting-receipt.ts
import { z } from 'zod';

export const SubcontractingReceiptSchema = z.object({
  id: z.string(),
  owner: z.string().optional(),
  creation: z.string().optional(),
  modified: z.string().optional(),
  modified_by: z.string().optional(),
  docstatus: z.number().default(0),
  title: z.string().optional().default('{supplier_name}'),
  naming_series: z.enum(['MAT-SCR-.YYYY.-', 'MAT-SCR-RET-.YYYY.-']),
  supplier: z.string(),
  supplier_name: z.string().optional(),
  supplier_delivery_note: z.string().optional(),
  company: z.string(),
  posting_date: z.string().default('Today'),
  posting_time: z.string().default('Now'),
  set_posting_time: z.boolean().optional().default(false),
  is_return: z.boolean().optional().default(false),
  return_against: z.string().optional(),
  cost_center: z.string().optional(),
  project: z.string().optional(),
  set_warehouse: z.string().optional(),
  rejected_warehouse: z.string().optional(),
  supplier_warehouse: z.string().optional(),
  items: z.array(z.unknown()),
  total_qty: z.number().optional(),
  total: z.number().optional(),
  supplied_items: z.array(z.unknown()).optional(),
  in_words: z.string().max(240).optional(),
  bill_no: z.string().optional(),
  bill_date: z.string().optional(),
  supplier_address: z.string().optional(),
  contact_person: z.string().optional(),
  address_display: z.string().optional(),
  contact_display: z.string().optional(),
  contact_mobile: z.string().optional(),
  contact_email: z.string().optional(),
  shipping_address: z.string().optional(),
  shipping_address_display: z.string().optional(),
  billing_address: z.string().optional(),
  billing_address_display: z.string().optional(),
  distribute_additional_costs_based_on: z.enum(['Qty', 'Amount']).optional().default('Qty'),
  additional_costs: z.array(z.unknown()).optional(),
  total_additional_costs: z.number().optional(),
  amended_from: z.string().optional(),
  range: z.string().optional(),
  represents_company: z.string().optional(),
  status: z.enum(['Draft', 'Completed', 'Return', 'Return Issued', 'Cancelled', 'Closed']).default('Draft'),
  per_returned: z.number().optional(),
  auto_repeat: z.string().optional(),
  letter_head: z.string().optional(),
  language: z.string().optional(),
  instructions: z.string().optional(),
  select_print_heading: z.string().optional(),
  other_details: z.string().optional(),
  remarks: z.string().optional(),
  transporter_name: z.string().optional(),
  lr_no: z.string().optional(),
  lr_date: z.string().optional(),
});

export type SubcontractingReceipt = z.infer<typeof SubcontractingReceiptSchema>;

export const SubcontractingReceiptInsertSchema = SubcontractingReceiptSchema.omit({ id: true, owner: true, creation: true, modified: true, modified_by: true });
export type SubcontractingReceiptInsert = z.infer<typeof SubcontractingReceiptInsertSchema>;